External input (#727)
* Implement get_external_input()

* Move external input/ output to separate class - otherwise the unit tests fail

* Add DocString and ValueError
