+++ pyiron/base/project/generic.py
@@ -41,6 +41,7 @@ from pyiron.base.server.queuestatus import (
     queue_enable_reservation,
     queue_check_job_is_waiting_or_running,
 )
+from pyiron.base.job.external import Notebook
 
 """
 The project object is the central import point of pyiron - all other objects can be created from this one
@@ -630,6 +631,22 @@ class Project(ProjectPath):
             element_lst=element_lst,
         )
         return df["status"].value_counts()
+
+    @staticmethod
+    def get_external_input():
+        """
+        Get external input either from the HDF5 file of the ScriptJob object which executes the Jupyter notebook
+        or from an input.json file located in the same directory as the Jupyter notebook. 
+        
+        Returns:
+            dict: Dictionary with external input
+        """
+        inputdict = Notebook.get_custom_dict()
+        if inputdict is None:
+            raise ValueError("No input found, either there is an issue with your ScriptJob, " + 
+                             "or your input.json file is not located in the same directory " +
+                             "as your Jupyter Notebook.")
+        return inputdict
 
     def keys(self):
         """
